+++ braindecode/models/attentionbasenet.py
@@ -175,7 +175,7 @@ class AttentionBaseNet(EEGModuleMixin, nn.Module):
                 f"n_times ({self.n_times}) is smaller than the minimum required "
                 f"({min_n_times_required}) for the current model parameters configuration. "
                 "Adjusting parameters to ensure compatibility."
-                "reducing the kernel, pooling, and stride sizes accordingly.\n"
+                "Reducing the kernel, pooling, and stride sizes accordingly.\n"
                 "Scaling factor: {:.2f}".format(scaling_factor),
                 UserWarning,
             )