+++ docs/whats_new.rst
@@ -25,6 +25,7 @@ Enhancements
 
 Bugs
 ~~~~
+- Fix initialization of :class:`braindecode.models.Deep4Net` (:gh:`752` by `Bruno Aristimunha`_)
 
 API changes
 ~~~~~~~~~~~