+++ docs/whats_new.rst
@@ -29,7 +29,7 @@ Bugs
 - Fix initialization of :class:`braindecode.models.Deep4Net` (:gh:`752` by `Bruno Aristimunha`_)
 - Fix initialization of :class:`braindecode.models.ATCNet` (:gh:`754` by `Bruno Aristimunha`_)
 - Fix Labram model for small input (:gh:`757` by `Bruno Aristimunha`_)
-- Fix the CTNet parameters to make the embedding configuration easy (:gh:`758` by `Bruno Aristimunha`_)
+- Fix the CTNet parameters to make the embedding configuration easy (:gh:`760` by `Bruno Aristimunha`_)
 
 API changes
 ~~~~~~~~~~~